Rewrite CSI Plugin stale node patch test case
